Move deinterleave_shuffle into module/core
